Splitting the cpp code generator out of the rest of the generators
This probably should have been how it was done originally, but it was easier to just make the code generator a part of the existing generator. But now the generators are over a thousand lines long so lets just start splitting them out to better guarentee low coupling.
